+++ src/Entities/Collection.php
@@ -44,6 +44,31 @@ class Collection extends NestedCollection
         return $this;
     }
 
+    /**
+     * Load the fields of a belongs to or belongs to many entity relation
+     *
+     * @param  string $relation
+     * @param  mixed  $fields  Single field name string or array of field names
+     * @return $this
+     */
+    public function loadRelationCanonicalPath($relation)
+    {
+        if ($this->isEmpty()) return $this;
+
+        if ( ! key_exists($relation, $this->first()->getAttributes())) {
+            $this->loadFields($relation);
+        }
+
+        $relations = $this->pluck($relation)->filter();
+        if ( ! $relations->first() instanceof Entity) {
+            $relations = $relations->flatten();
+        }
+
+        (new static($relations->unique()))->loadCanonicalPath();
+
+        return $this;
+    }
+
     public function injectValues()
     {
         foreach ($this->items as $item) {